+++ app/src/main/kotlin/ac/mdiq/podcini/storage/model/Episode.kt
@@ -9,7 +9,6 @@ import ac.mdiq.podcini.net.feed.parser.media.vorbis.VorbisCommentReaderException
 import ac.mdiq.podcini.net.utils.NetworkUtils.isImageDownloadAllowed
 import ac.mdiq.podcini.storage.database.getFeed
 import ac.mdiq.podcini.storage.database.upsert
-import ac.mdiq.podcini.storage.database.upsertBlk
 import ac.mdiq.podcini.storage.model.Feed.Companion.TAG_SEPARATOR
 import ac.mdiq.podcini.storage.specs.EpisodeState
 import ac.mdiq.podcini.storage.specs.MediaType
@@ -29,6 +28,7 @@ import ac.mdiq.podcini.utils.Logt
 import ac.mdiq.podcini.utils.fullDateTimeString
 import android.content.ContentResolver
 import android.content.Context
+import android.media.MediaMetadataRetriever
 import android.net.Uri
 import android.provider.OpenableColumns
 import android.webkit.URLUtil.guessFileName
@@ -55,7 +55,6 @@ import org.apache.commons.io.FilenameUtils.getExtension
 import org.apache.commons.io.input.CountingInputStream
 import org.apache.commons.lang3.builder.ToStringBuilder
 import org.apache.commons.lang3.builder.ToStringStyle
-import wseemann.media.FFmpegMediaMetadataRetriever
 import java.io.BufferedInputStream
 import java.io.File
 import java.io.FileInputStream
@@ -883,17 +882,45 @@ class Episode : RealmObject {
         return true
     }
 
+//    fun checkEmbeddedPicture(persist: Boolean = true) {
+//        if (!localFileAvailable()) hasEmbeddedPicture = false
+//        else {
+//            var retriever: FFmpegMediaMetadataRetriever? = null
+//            try {
+//                retriever = FFmpegMediaMetadataRetriever()
+//                retriever.setDataSource(fileUrl?.toUri().toString())
+//                hasEmbeddedPicture = (retriever.embeddedPicture != null)
+//            } catch (e: Exception) { Logs(TAG, e, "checkEmbeddedPicture failed.") } finally { retriever?.release() }
+//        }
+//        if (persist) upsertBlk(this) {}
+//    }
+
     fun checkEmbeddedPicture(persist: Boolean = true) {
         if (!localFileAvailable()) hasEmbeddedPicture = false
         else {
-            var retriever: FFmpegMediaMetadataRetriever? = null
-            try {
-                retriever = FFmpegMediaMetadataRetriever()
-                retriever.setDataSource(fileUrl?.toUri().toString())
-                hasEmbeddedPicture = (retriever.embeddedPicture != null)
-            } catch (e: Exception) { Logs(TAG, e, "checkEmbeddedPicture failed.") } finally { retriever?.release() }
+            // TODO: what to do with this
+//            try {
+//                MediaMetadataRetrieverCompat().use { mmr ->
+//                    mmr.setDataSource(getAppContext(), Uri.parse(fileUrl))
+//                    val image = mmr.embeddedPicture
+//                    hasEmbeddedPicture = image != null
+//                }
+//            } catch (e: Exception) {
+//                Logs(TAG, e)
+//                hasEmbeddedPicture = false
+//            }
         }
-        if (persist) upsertBlk(this) {}
+        // TODO
+//        if (persist && episode != null) upsertBlk(episode!!) {}
+    }
+
+    /**
+     * On SDK<29, this class does not have a close method yet, so the app crashes when using try-with-resources.
+     */
+    class MediaMetadataRetrieverCompat : MediaMetadataRetriever(), AutoCloseable {
+        override fun close() {
+            try { release() } catch (e: IOException) { Logs(TAG, e, "MediaMetadataRetriever failed") }
+        }
     }
 
     // above from EpisodeMedia